+++ src/components/10-atoms/spinner/e2e.js
@@ -5,6 +5,6 @@ test.describe('spinner', () => {
   test('should render spinner', async ({ page }) => {
     await page.goto(fixtureURL('components-spinner--spinner'));
 
-    await expect(page.locator('axa-spinner')).toBeVisible();
+    await expect(page.locator('.a-spinner')).toBeVisible();
   });
 });