+++ src/other/pages/guides/getting-started/story.jsx
@@ -24,10 +24,10 @@ export const GettingStarted = () =>
     <div className="accessory-story-wrapper">
       <style>{styles}</style>
       <div className="accessory-story-content">
-        <AXAHeading rank="1" secondary>
+        <AXAHeading size="1" secondary>
           Getting started
         </AXAHeading>
-        <AXAText variant="size-2">
+        <AXAText size="2">
           The AXA Design System is a toolbox of resources to create beautiful
           user interfaces that are consistent with AXA brand guidelines,
           principles, and best practices. Instead of focusing on pixels,
@@ -36,18 +36,18 @@ export const GettingStarted = () =>
         </AXAText>
         <header className="getting-started__header">
           <p className="getting-started__subtitle">A perfect start for a</p>
-          <AXAHeading rank="2" secondary>
+          <AXAHeading size="2" secondary>
             Developer
           </AXAHeading>
         </header>
-        <AXAText variant="size-2">
+        <AXAText size="2">
           The library provides front-end developers & engineers with a
           collection of reusable Web Components to build websites and user
           interfaces that are aligned with AXA brand guidelines. Adopting the
           AXA components library empowers you to use consistent markup, styling
           and behavior in both prototype and production work.
         </AXAText>
-        <AXAHeading rank="3">Install and use components</AXAHeading>
+        <AXAHeading size="3">Install and use components</AXAHeading>
         <p className="getting-started__dependency-wrapper">
           AXA&rsquo;s Design System is made up of multiple web components and
           tools, which you can import one by one. All you need to do is install
@@ -117,7 +117,7 @@ export default AXAButtonReact;`}
   I am a Button
 </AXAButtonReact>`}
         </ReactSyntaxHighlighter>
-        <AXAHeading rank="3">Read more about the Pattern Library</AXAHeading>
+        <AXAHeading size="3">Read more about the Pattern Library</AXAHeading>
         <PLCallout
           className="getting-started__callout-first"
           icon="assets/github-mark.png"
@@ -149,12 +149,12 @@ export default AXAButtonReact;`}
         />
         <header className="getting-started__header">
           <p className="getting-started__subtitle">A perfect start for a</p>
-          <AXAHeading rank="2" secondary>
+          <AXAHeading size="2" secondary>
             Designer
           </AXAHeading>
         </header>
         <div className="getting-started__last-text">
-          <AXAText variant="size-2">
+          <AXAText size="2">
             It is our aim to offer our customers safety and trust when using our
             services. To achieve this, we rely on uniform visual communication
             and use existing and tested interactions.
@@ -167,9 +167,9 @@ export default AXAButtonReact;`}
           link="https://www.figma.com/file/6zurYk3bJpzUg0H2THSxGF/AXA-UI-Kit?node-id=0%3A8208"
           linkText="READ MORE"
         />
-        <AXAHeading rank="3">Our development guide</AXAHeading>
+        <AXAHeading size="3">Our development guide</AXAHeading>
         <div className="getting-started__last-text">
-          <AXAText variant="size-2">
+          <AXAText size="2">
             In order for us to screen your application for go-live, the
             following requirements have to be considered during product
             development: